// src/store/main/device/imei/types.ts
/*
 * @Author: korealu
 * @Date: 2022-02-16 16:53:07
 * @LastEditors: korealu 643949593@qq.com
 * @LastEditTime: 2022-05-06 14:40:14
 * @Description: file content
 * @FilePath: /pofi-admin/src/store/main/device/imei/types.ts
 */
export interface IUserImeiType {
  imeiList: any[]
  imeiCount: number
  queryInfo: any
}